finished separating intets up
