+++ lambda/py/lambda_function.py
@@ -1,5 +1,5 @@
 """
-Skill to help with crafting options in Terraria
+lambda_function deals with the launch handler and the skill building.
 """
 #--------------------------------------------------------------------------
 # Python Libraries
@@ -27,6 +27,7 @@ from skill import HelpIntentHandler, CancelOrStopIntentHandler, RepeatIntentHand
     FallbackIntentHandler, SessionEndedRequestHandler, CatchAllExceptionHandler, \
     CacheSpeechForRepeatInterceptor, LocalizationInterceptor
 from skill import RandomItemIntentHandler
+from skill import RecipeIntentHandler
 
 # Set Logging 
 logger = logging.getLogger(__name__)
@@ -62,63 +63,6 @@ class LaunchRequestHandler(AbstractRequestHandler):
         handler_input.response_builder.speak(speech).set_card(
             SimpleCard(data.SKILL_NAME, "Terraria Tool")
         ).ask(reprompt)
-
-        return handler_input.response_builder.response
-
-
-class RecipeIntentHandler(AbstractRequestHandler):
-    """
-    Handler for getting the recipe for an item
-
-    Parameters:
-    AbstractRequestHandler (obj): Amazon Request Handler object
-
-    Returns: Amazon request with the item the user requested
-    """
-    def can_handle(self, handler_input):
-        # type: (HandlerInput) -> bool
-        return is_intent_name("RecipeIntent")(handler_input)
-
-    def handle(self, handler_input):
-        # type: (HandlerInput) -> Response
-        logger.info("In RecipeIntentHandler")
-
-        _ = handler_input.attributes_manager.request_attributes["_"]
-
-        try:
-            item_name = handler_input.request_envelope.request.intent.slots["Item"].value.lower()
-        except AttributeError:
-            logger.info("Could not resolve item name")
-            item_name = None
-
-        card_title = _(data.DISPLAY_CARD_TITLE).format(_(data.SKILL_NAME), item_name)
-
-        list_of_items = data.ITEMS
-
-        # I was dumb and didn't put all 3000 items as lowercase...so here we are
-        #TODO: Eventually fix the dictionary, but get this working first
-        list_of_items = {k.lower(): v for k, v in list_of_items.items()}
-
-        if item_name in list_of_items:
-            recipe = list_of_items[item_name]
-            # session_attributes['speech'] = recipe
-            handler_input.response_builder.speak(recipe).set_card(
-                SimpleCard(card_title, recipe)).set_should_end_session(True)
-        else:
-            speech = _(data.ITEM_NOT_FOUND_MESSAGE)
-            reprompt = _(data.ITEM_NOT_FOUND_REPROMPT)
-            if item_name:
-                speech += _(data.ITEM_NOT_FOUND_WITH_ITEM_NAME).format(
-                    item_name)
-            else:
-                speech += _(data.ITEM_NOT_FOUND_WITHOUT_ITEM_NAME)
-            speech += reprompt
-
-            handler_input.response_builder.speak(speech).ask(
-                reprompt)
-
-        #TODO: Is there a way to put in here "would you like to know another item" 
-        # instead of having to open the skill multiple times?
 
         return handler_input.response_builder.response
